practice_3+4/input/pyproject.toml
[project]
name = "mcp-validation-debugging-practice"
version = "0.1.0"
description = "MCP server validation and debugging practice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "httpx>=0.27.0",
    "mcp>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
